Create digitRecognition.py
This is currently an ANN that is trained off the MNIST dataset, which is built into PyTorch. This code will not be part of the actual application, as this is the code to train and test the model. I need to figure out how to export the model once it is completed. Within the next week, I will convert this ANN to a CNN, as it is more suitable for our needs.